[#11] refactor: extract auth resolution into src/auth.ts
