+++ src/auth.ts
@@ -1,0 +1,30 @@
+import * as fs from "node:fs/promises"
+import * as path from "node:path"
+import { xdgData } from "xdg-basedir"
+import type { OpenAIAuth } from "./types"
+
+// Mirrors OpenCode's auth resolution: OPENCODE_AUTH_CONTENT overrides $XDG_DATA_HOME/opencode/auth.json.
+// The Auth service is not exposed to external plugins, so this reproduces the rules directly.
+async function loadAuthData(): Promise<Record<string, unknown>> {
+  if (process.env.OPENCODE_AUTH_CONTENT) {
+    return JSON.parse(process.env.OPENCODE_AUTH_CONTENT) as Record<string, unknown>
+  }
+  if (!xdgData) {
+    throw new Error("could not determine XDG data directory")
+  }
+  const raw = await fs.readFile(path.join(xdgData, "opencode", "auth.json"), "utf-8")
+  return JSON.parse(raw) as Record<string, unknown>
+}
+
+export async function loadOpenAIAuth(): Promise<OpenAIAuth | undefined> {
+  try {
+    const data = await loadAuthData()
+    const entry = data.openai as Partial<OpenAIAuth> | undefined
+    if (entry?.type === "oauth" && typeof entry.access === "string") {
+      return entry as OpenAIAuth
+    }
+  } catch {
+    return undefined
+  }
+  return undefined
+}
