+++ src/index.ts
@@ -4,7 +4,7 @@ import type { Hooks, Plugin, PluginInput, PluginModule } from "@opencode-ai/plug
 import { tool } from "@opencode-ai/plugin"
 import { EventSourceParserStream } from "eventsource-parser/stream"
 import { fileTypeFromBuffer } from "file-type"
-import { xdgData } from "xdg-basedir"
+import { loadOpenAIAuth } from "./auth"
 import { saveGeneratedImage } from "./output-image"
 import type { GenerateArgs, OpenAIAuth } from "./types"
 
@@ -25,32 +25,6 @@ async function readImageAsDataUrl(filePath: string, ctxDir: string): Promise<str
     throw new Error(`unsupported image file type: ${abs}`)
   }
   return `data:${detected.mime};base64,${buf.toString("base64")}`
-}
-
-// Mirrors OpenCode's auth resolution: OPENCODE_AUTH_CONTENT overrides $XDG_DATA_HOME/opencode/auth.json.
-// The Auth service is not exposed to external plugins, so this reproduces the rules directly.
-async function loadAuthData(): Promise<Record<string, unknown>> {
-  if (process.env.OPENCODE_AUTH_CONTENT) {
-    return JSON.parse(process.env.OPENCODE_AUTH_CONTENT) as Record<string, unknown>
-  }
-  if (!xdgData) {
-    throw new Error("could not determine XDG data directory")
-  }
-  const raw = await fs.readFile(path.join(xdgData, "opencode", "auth.json"), "utf-8")
-  return JSON.parse(raw) as Record<string, unknown>
-}
-
-async function loadOpenAIAuth(): Promise<OpenAIAuth | undefined> {
-  try {
-    const data = await loadAuthData()
-    const entry = data.openai as Partial<OpenAIAuth> | undefined
-    if (entry?.type === "oauth" && typeof entry.access === "string") {
-      return entry as OpenAIAuth
-    }
-  } catch {
-    return undefined
-  }
-  return undefined
 }
 
 type CodexSSEEvent = {
